Prepare code to reuse
