--- app/threads/client/components/hooks/useUserRoom.js
@@ -1,4 +1,0 @@
-import { useReactiveValue } from '../../../../../client/hooks/useReactiveValue';
-import { Rooms } from '../../../../models/client';
-
-export const useUserRoom = (rid, fields) => useReactiveValue(() => Rooms.findOne({ _id: rid }, { fields }), [rid, fields]);
