--- app/threads/client/components/hooks/useUserSubscription.js
@@ -1,4 +1,0 @@
-import { useReactiveValue } from '../../../../../client/hooks/useReactiveValue';
-import { Subscriptions } from '../../../../models/client';
-
-export const useUserSubscription = (rid, fields) => useReactiveValue(() => Subscriptions.findOne({ rid }, { fields }), [rid, fields]);
