+++ client/Channel/hooks/useUserSubscription.js
@@ -1,0 +1,4 @@
+import { useReactiveValue } from '../../hooks/useReactiveValue';
+import { Subscriptions } from '../../../app/models/client';
+
+export const useUserSubscription = (rid, fields) => useReactiveValue(() => Subscriptions.findOne({ rid }, { fields }), [rid, fields]);
